my first cask-- program is of an fps
This is my first cask, so I can only hope that it works properly for everyone. If there is something wrong, please feel free to contact me.

There are also nightly builds that are supplied, however, as well as needing to go under another cask directory, the builds themselves require compilation, as they are only the binaries- I want to learn to walk before I can run ;-P. However, I hope to be able to wrap my mind around the problem some time soon.

Tesseract is a free-to-play FPS, created for commercial use, but has since received little recognition, as well as having very few developers, leading to an under-developed, under-recognised game that has great potential.